+++ helpers.h
@@ -17,11 +17,24 @@ long long int karmarkar_karp(long long int * A, int n);
 class Return
 {
 public:
-	std::vector<double> times;
+	std::vector<float> times;
 	std::vector<long long int> resids;
-	void add(double t, long long int r);
+	void add(float t, long long int r);
 	void print_times();
 	void print_resids();
 };
+
+// class Solution
+// {
+// private:
+// 	long long int * A;
+// public:
+// 	std::std::vector<int> P;
+// 	long long int resid;
+// 	Solution(long long int *A, int n);
+// 	void randomize();
+// 	void neighbor();
+// 	void update();
+// };
 
 #endif //CS124PA3_KK_H
